pyproject.toml
[tool.poetry]
name = "stitchtly-chatbot"
version = "0.1.0"
description = ""
authors = ["jorgechvz <cha22060@byui.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
llama-index-llms-openai = "^0.3.1"
llama-index = "^0.12.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
